// src/server/actions/create-agency/schema.ts
import { z } from "zod";

export const CreateAdmin = z.object({
  id: z.string().optional(),
  customerId: z.string().optional(),
  name: z.string().min(2, { message: "Agency name must be atleast 2 chars." }),
  companyEmail: z.string().min(1),
  companyPhone: z.string().min(1),
  whiteLabel: z.boolean(),
  address: z.string().min(1),
  city: z.string().min(1),
  zipCode: z.string().min(1),
  state: z.string().min(1),
  country: z.string().min(1),
  agencyLogo: z.string().min(1),
});